+++ NEWS.md
@@ -3,7 +3,7 @@ __kaiaulu 0.0.0.9700 (in development)__
 
 ### NEW FEATURES
 
- * New function `engagement_communication` added to `R/smells.R` uses a 90-day rollig window to check, for each window, how many times an author communicated via a message. [#360](https://github.com/sailuh/kaiaulu/issues/360)
+ * New functions `engagement_communication` and `engagement_sentiment` added to `R/metric.R`. The purpose is to use a 90-day rollig window to check, for each window, how many times an author communicated via a message and determine the average sentiment value of an author's messages. [#360](https://github.com/sailuh/kaiaulu/issues/360)
  *  A new capability to export events for process mining has been added. [#301](https://github.com/sailuh/kaiaulu/issues/301)
  * `exec/ghevents.R` has been added. It is a executable CLI (command-line interface) to download and parse Github Issue Events from outside Kaiaulu.
  * All GitHub Pull Request Comments are able to be downloaded in the Pull Request Comments notebook. [342](https://github.com/sailuh/kaiaulu/issues/342)